Use an explicit prefilter when a hbbft transaction set has been agreed
Prior to this change, we had 2 calls to a sidecar transaction filter
operation, one of them was done as soon as the block's contents were
agreed on and one was done at the start of a new round, if the prior
filtering had not been done.

The prefilter had a problem that it didn't tell the sidecar what
transactions were pending. This led to the sidecar rejecting all
transactions that depended on the transactions about to appear in the
next block.

This change adds an explicit prefilter that absorbs the valid
transactions from the pending block and then re-evaluates all the
buffered transactions against that. This means that transactions that
depend on the tranactions in the pending block will no longer be
filtered out.

Additionally, this change seperates pending transactions in the next
block from invalid transactions in the return from miner:create_block so
we don't bother trying to absorb invalid transactions in the prefilter.