Treat onboarding key as token, clean up info CLI.
Change the code in`miner_keys` so that the onboarding "key" is
treated more like an opaque token rather than a string of any particular
format. It used to be the case that the onboarding key was actually a
public key, encoded in base58 format, but with the advent of DIY and,
later, HIP-19, this is no longer the case. The code base was starting to
bend in strange ways to support this old notion, particularly in
`print_keys/1`, where the function had to duplicate the logic of
determining the source of the onboarding key just to figure out how to
convert it into a string. By forcing it to be a string when it is loaded
from its ultimate source, we remove this unusual coupling and hopefully,
squash a Dialyzer warning that must have existed for some time.

Polish up the "info onboarding" CLI handler by making it a bit more
idiomatic Erlang, and allowing the user to provide _any_ onboarding key
for the onboarding key query if desired, not just the key in this miner.