Remove key generation from get_public_key
get_public_key had the nasty sideeffect of generating keys when they were not found in a slot.

In the new Maker world the public key is intended to be the onbarding key which means slot 15 was empty, but the code was still generating a key there which was then wrongly used as the onboarding key. This fix will correct that, and safe because slot 0 is already populated by a key by gateway_mfr

Old hotspots that were not provisioned with a slot 0 key will have old miner_keys code that _will_ generate a key in slot 0 before they get an OTA