Fix BBA vote default to be everybody completed
Currently if we cannot agree on a BBA result vector we default to the
empty binary (all 0s). This results in *all* nodes taking a BBA penalty
hit. This is incorrect and this change defaults to the assumption that
everyone completed their BBA so that we do not penalize the entire
group.

This change may cause a "split block" but as long as it merges after the
time based autoskip it should be safe once enough nodes have upgraded.