Start next round early if node does not have its proposal in
If a node does not get its proposal into a round have it start the next
round early to get a jump on things. This should help with cases where
the node is behind and, because it has to validate all the blocks it
didn't sign, can't catch up for the whole epoch.