Fix an issue which was preventing use of RSSIS when available.
Fix an issue which was preventing the use of the "rssis" signal report
on a packet, even when it was present and desired. RSSIS or "RSSI of signal"
is an additional packet meta-data item provided by the SX1302 (and other)
packet forwarders, and when present, it more accurately reflects the power
level at which a packet was received. In contrast, the "rssi" value on such
forwarders reflects the "RSSI of channel", which is closer to the RSSI of the
channel noise floor, not the packet that was received.

Since "rssis" is a more accurate gauge of reception power, it is preferred
by PoC v11, but wasn't actually being used because of this issue.

Based on an original commit by Amir Haleem.